+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Hebrew Translation (thanks to Ahiel and Natan)
 - Dutch Translation (thanks to all contributors)
 - Setting a preference that doesn't actually change the value won't have any effects (such as MQTT reconnecting) (#1875)
+- Status messages contain the app version and flavour
 
 ### Bug fixes